company/requirements.txt
asgiref==3.3.1
autopep8==1.5.5
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
curl==0.0.1
decorators==2.0.7
Django==3.1.7
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.1
django-filter==1.1.0
django-js-asset==1.2.2
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-filters==0.11.1
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
Markdown==3.3.4
MarkupSafe==1.1.1
mock==4.0.3
openapi-codec==1.3.2
packaging==20.9
Pillow==8.1.2
pycodestyle==2.6.0
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
